+++ share/pot/create.sh
@@ -3,7 +3,7 @@
 create-help()
 {
 	echo "pot create [-hv] -p potname [-i ipaddr] [-l lvl] [-f flavour|-F]"
-	echo '  [-b base | -P basepot ] [-d dns]'
+	echo '  [-b base | -P basepot ] [-d dns] [-t type]'
 	echo '  -h print this help'
 	echo '  -v verbose'
 	echo '  -p potname : the pot name (mandatory)'
@@ -15,33 +15,59 @@ create-help()
 	echo '  -d dns : one between inherit(default) or pot'
 	echo '  -f flavour : flavour to be used'
 	echo '  -F : no default flavour is used'
+	echo '  -t type: single or multi (default multi)'
+    echo '         single: the pot is based on a unique ZFS dataset'	
+    echo '         multi: the pot is composed by a classical collection of 3 ZFS dataset'	
 }
 
 # $1 pot name
-# $2 base name
+# $2 type
 # $3 level
-# $4 pot-base name
+# $4 base name
+# $5 pot-base name
 _cj_zfs()
 {
-	local _pname _base _potbase _jdset _snap _custom _dset
+	local _pname _base _type _potbase _jdset _snap _dset
 	_pname=$1
-	_base=$2
+	_type=$2
 	_lvl=$3
-	_potbase=$4
+	_base=$4
+	_potbase=$5
 	_jdset=${POT_ZFS_ROOT}/jails/$_pname
 	# Create the main jail zfs dataset
-	if ! _zfs_dataset_valid $_jdset ; then
-		zfs create $_jdset
+	if ! _zfs_dataset_valid "$_jdset" ; then
+		zfs create "$_jdset"
 	else
 		_info "$_jdset exists already"
 	fi
+	if [ "$_type" = "single" ]; then
+		if [ -z "$_potbase" ]; then
+			# create an empty dataset
+			zfs create "$_jdset/m"
+			# create the minimum needed tree
+			mkdir -p "${POT_FS_ROOT}/jails/$_pname/m/tmp"
+			mkdir -p "${POT_FS_ROOT}/jails/$_pname/m/dev"
+		else
+			# clone the last snapshot of _potbase
+			_dset=${POT_ZFS_ROOT}/jails/$_potbase/m
+			_snap=$(_zfs_last_snap "$_dset")
+			if [ -n "$_snap" ]; then
+				_debug "Clone zfs snapshot $_dset@$_snap"
+				zfs clone -o mountpoint="${POT_FS_ROOT}/jails/$_pname/m" "$_dset/@$_snap" "$_jdset/m"
+			else
+				# TODO - autofix
+				_error "no snapshot found for $_dset/m"
+				return 1 # false
+			fi
+		fi
+		return 0
 	# Create the root mountpoint
-	if [ ! -d "${POT_FS_ROOT}/jails/$_pname/m" ]; then
-		mkdir -p ${POT_FS_ROOT}/jails/$_pname/m
+	elif [ ! -d "${POT_FS_ROOT}/jails/$_pname/m" ]; then
+		mkdir -p "${POT_FS_ROOT}/jails/$_pname/m"
 	fi
 
 	# lvl 0 images mount directly usr.local and custom
-	if [ $_lvl -eq 0 ]; then
+	if [ "$_lvl" = "0" ]; then
 		return 0 # true
 	fi
 
@@ -96,10 +122,11 @@ _cj_zfs()
 # $4 static_ip
 # $5 level
 # $6 dns
-# $7 pot-base name
+# $7 type
+# $8 pot-base name
 _cj_conf()
 {
-	local _pname _base _ip _staticip _lvl _jdir _bdir _potbase _dns
+	local _pname _base _ip _staticip _lvl _jdir _bdir _potbase _dns _type
 	local _pblvl _pbpb
 	local _jdset _bdset _pbdset _baseos
 	_pname=$1
@@ -108,7 +135,8 @@ _cj_conf()
 	_staticip=$4
 	_lvl=$5
 	_dns=$6
-	_potbase=$7
+	_type=$7
+	_potbase=$8
 	_jdir=${POT_FS_ROOT}/jails/$_pname
 	_bdir=${POT_FS_ROOT}/bases/$_base
 
@@ -124,6 +152,7 @@ _cj_conf()
 		mkdir -p $_jdir/conf
 	fi
 	(
+	if [ "$_type" = "multi" ]; then
 		case $_lvl in
 		0)
 			echo "$_bdset ${_jdir}/m"
@@ -146,13 +175,16 @@ _cj_conf()
 			echo "$_jdset/custom ${_jdir}/m/opt/custom zfs-remount"
 			;;
 		esac
+	fi
 	) > $_jdir/conf/fscomp.conf
 	(
 		_baseos=$( cat $_bdir/.osrelease )
 		echo "pot.level=${_lvl}"
+		echo "pot.type=${_type}"
 		echo "pot.base=${_base}"
 		echo "pot.potbase=${_potbase}"
 		echo "pot.dns=${_dns}"
+		echo "pot.cmd=sh /etc/rc"
 		echo "host.hostname=\"${_pname}.$( hostname )\""
 		echo "osrelease=\"${_baseos}-RELEASE\""
 		if [ "$_ip" = "inherit" ]; then
@@ -167,15 +199,15 @@ _cj_conf()
 				echo "vnet=true"
 			fi
 		fi
-		if [ "${_dns}" == "pot" ]; then
+		if [ "${_dns}" = "pot" ]; then
 			echo "pot.depend=${POT_DNS_NAME}"
 		fi
 	) > $_jdir/conf/pot.conf
-	if [ $_lvl -eq 2 ]; then
+	if [ "$_lvl" -eq 2 ]; then
 		if [ $_pblvl -eq 1 ]; then
 			# CHANGE the potbase usr.local to be not zfs-remount
 			# Add an info here would be nice
-			if [ -w ${POT_FS_ROOT}/jails/$_potbase/conf/fscomp.conf ]; then
+			if [ -w "${POT_FS_ROOT}/jails/$_potbase/conf/fscomp.conf" ]; then
 				_info "${POT_FS_ROOT}/jails/$_potbase/conf/fscomp.conf fix (${POT_FS_ROOT}/jails/$_potbase/m/usr/local zfs-remount)"
 				${SED} -i '' s%${POT_FS_ROOT}/jails/$_potbase/m/usr/local\ zfs-remount%${POT_FS_ROOT}/jails/$_potbase/m/usr/local% ${POT_FS_ROOT}/jails/$_potbase/conf/fscomp.conf
 			else
@@ -184,28 +216,30 @@ _cj_conf()
 		fi
 	fi
 	# disable some cron jobs, not relevant in a jail
-	if [ $_lvl -ne 0 ]; then
+	if [ "$_lvl" -ne 0 ]; then
 		${SED} -i '' 's/^.*save-entropy$/# &/g' "${_jdir}/custom/etc/crontab"
 		${SED} -i '' 's/^.*adjkerntz.*$/# &/g' "${_jdir}/custom/etc/crontab"
 	fi
 
-	# add remote syslogd capability, if not inherit
-	if [ "$_ip" != "inherit" ]; then
-		# configure syslog in the pot
-		${SED} -i '' 's%^[^#].*/var/log.*$%# &%g' "${_jdir}/custom/etc/syslog.conf"
-		echo "*.*  @${POT_GATEWAY}:514" > "${_jdir}/custom/etc/syslog.d/pot.conf"
-		sysrc -f "${_jdir}/custom/etc/rc.conf" "syslogd_flags=-vv -s -b $_ip"
-		# configure syslogd in the host
-		(
-			echo +"$_ip"
-			echo '*.*		'"/var/log/pot/${_pname}.log"
-		) > /usr/local/etc/syslog.d/"${_pname}".conf
-		touch /var/log/pot/"${_pname}".log
-		(
-			echo "# log rotation for pot ${_pname}"
-			echo "/var/log/pot/${_pname}.log 644 7 * @T00 CX"
-		) > /usr/local/etc/newsyslog.conf.d/"${_pname}".conf
-		service syslogd reload
+	if [ "$_type" = "multi" ]; then
+		# add remote syslogd capability, if not inherit
+		if [ "$_ip" != "inherit" ]; then
+			# configure syslog in the pot
+			${SED} -i '' 's%^[^#].*/var/log.*$%# &%g' "${_jdir}/custom/etc/syslog.conf"
+			echo "*.*  @${POT_GATEWAY}:514" > "${_jdir}/custom/etc/syslog.d/pot.conf"
+			sysrc -f "${_jdir}/custom/etc/rc.conf" "syslogd_flags=-vv -s -b $_ip"
+			# configure syslogd in the host
+			(
+				echo +"$_ip"
+				echo '*.*		'"/var/log/pot/${_pname}.log"
+			) > /usr/local/etc/syslog.d/"${_pname}".conf
+			touch /var/log/pot/"${_pname}".log
+			(
+				echo "# log rotation for pot ${_pname}"
+				echo "/var/log/pot/${_pname}.log 644 7 * @T00 CX"
+			) > /usr/local/etc/newsyslog.conf.d/"${_pname}".conf
+			service syslogd reload
+		fi
 	fi
 }
 
@@ -242,7 +276,7 @@ _cj_flv()
 pot-create()
 {
 	local _pname _ipaddr _lvl _base _flv _potbase
-	local _flv_default _dns _staticip
+	local _flv_default _dns _staticip _type
 	_pname=
 	_base=
 	_ipaddr=inherit
@@ -252,8 +286,8 @@ pot-create()
 	_flv_default="YES"
 	_dns=inherit
 	_staticip="NO"
-	args=$(getopt hvp:i:sl:b:f:P:Fd: $*)
-	if [ $? -ne 0 ]; then
+	_type="multi"
+	if ! args=$(getopt hvp:i:sl:b:f:P:Fd:t: "$@") ; then
 		create-help
 		${EXIT} 1
 	fi
@@ -282,6 +316,16 @@ pot-create()
 			;;
 		-l)
 			_lvl=$2
+			shift 2
+			;;
+		-t)
+			if [ "$2" = "multi" ] || [ "$2" = "single" ]; then
+				_type="$2"
+			else
+				_error "Type $2 not supported"
+				create-help
+				${EXIT} 1
+			fi
 			shift 2
 			;;
 		-b)
@@ -332,118 +376,151 @@ pot-create()
 	done
 
 	# check options consitency
-	case $_lvl in
-		0)
+	if [ "$_type" = "single" ]; then
+		_lvl=0
+		_flv_default="NO"
+		if [ -n "$_potbase" ]; then
+			if ! is_pot "$_potbase" quiet ; then
+				_error "pot $_potbase not found"
+				${EXIT} 1
+			fi
+			if [ "$( _get_pot_type "$_potbase" )" != "single" ]; then
+				_error "pot $_potbase has the wrong type, it as to be of type single"
+				${EXIT} 1
+			fi
 			if [ -z "$_base" ]; then
-				_error "level $_lvl needs option -b"
+				_base="$( _get_pot_base "$_potbase" )"
+			elif [ "$( _get_pot_base "$_potbase" )" != "$_base" ]; then
+				_error "-b $_base and -P $_potbase are not compatible"
 				create-help
 				${EXIT} 1
 			fi
-			if [ -n "$_potbase" ]; then
-				_error "-P option is not allowed with level $_lvl"
-				create-help
-				${EXIT} 1
-			fi
-			if ! _is_base "$_base" quiet ; then
-				_error "$_base is not a valid base"
-				create-help
-				${EXIT} 1
-			fi
-			;;
-		1)
-			if [ -z "$_base" -a -z "$_potbase" ]; then
+		else
+		   	if [ -z "$_base" ]; then
 				_error "at least one of -b and -P has to be used"
 				create-help
 				${EXIT} 1
 			fi
-			if [ -n "$_base" -a -n "$_potbase" ]; then
-				if [ "$( _get_pot_base $_potbase )" != "$_base" ]; then
-					_error "-b $_base and -P $_potbase are not compatible"
-					create-help
-					${EXIT} 1
-				fi
-				# TODO: an info or debug message che be showned
-			fi
-			if [ -n "$_potbase" ]; then
-				if ! _is_pot $_potbase ; then
-					_error "-P $_potbase : is not a pot"
-					create-help
-					${EXIT} 1
-				fi
-				if [ "$( _get_conf_var $_potbase pot.level )" != "1" ]; then
-					_error "-P $_potbase : it has to be of level 1"
-					create-help
-					${EXIT} 1
-				fi
-			fi
-			if [ -z "$_base" ]; then
-				_base=$( _get_pot_base $_potbase )
-				if [ -z "$_base" ]; then
-					_error "-P $potbase has no base??"
-					${EXIT} 1
-				fi
-				_debug "-P $_potbase induced -b $_base"
-			fi
 			if ! _is_base "$_base" quiet ; then
 				_error "$_base is not a valid base"
 				create-help
 				${EXIT} 1
 			fi
-			;;
-		2)
-			if [ -z "$_potbase" ]; then
-				_error "level $_lvl pots need another pot as reference"
-				create-help
-				${EXIT} 1
-			fi
-			if [ $( _get_conf_var $_potbase pot.level ) -lt 1 ]; then
-				_error "-P $_potbase : it has to be at least of level 1"
-				create-help
-				${EXIT} 1
-			fi
-			if ! _is_pot $_potbase ; then
-				_error "-P $_potbase : is not a pot"
-				create-help
-				${EXIT} 1
-			fi
-			if [ -n "$_base" ]; then
+		fi
+	else
+		case $_lvl in
+				0)
+				if [ -z "$_base" ]; then
+					_error "level $_lvl needs option -b"
+					create-help
+					${EXIT} 1
+				fi
+				if [ -n "$_potbase" ]; then
+					_error "-P option is not allowed with level $_lvl"
+					create-help
+					${EXIT} 1
+				fi
 				if ! _is_base "$_base" quiet ; then
 					_error "$_base is not a valid base"
 					create-help
 					${EXIT} 1
 				fi
-				if [ "$( _get_pot_base $_potbase )" != "$_base" ]; then
-					_error "-b $_base and -P $_potbase are not compatible"
-					${EXIT} 1
-				fi
-			else
-				_base=$( _get_pot_base $_potbase )
-				if [ -z "$_base" ]; then
-					_error "-P $potbase has no base??"
-					${EXIT} 1
-				fi
-				if ! _is_base "$_base" quiet ; then
-					_error "$_base (induced by the pot $_potbase) is not a valid base"
+				;;
+			1)
+				if [ -z "$_base" -a -z "$_potbase" ]; then
+					_error "at least one of -b and -P has to be used"
 					create-help
 					${EXIT} 1
 				fi
-			fi
-			;;
-		*)
-			_error "level $_lvl is not supported"
-			${EXIT} 1
-			;;
-	esac
+				if [ -n "$_base" -a -n "$_potbase" ]; then
+					if [ "$( _get_pot_base $_potbase )" != "$_base" ]; then
+						_error "-b $_base and -P $_potbase are not compatible"
+						create-help
+						${EXIT} 1
+					fi
+					# TODO: an info or debug message che be showned
+				fi
+				if [ -n "$_potbase" ]; then
+					if ! _is_pot $_potbase ; then
+						_error "-P $_potbase : is not a pot"
+						create-help
+						${EXIT} 1
+					fi
+					if [ "$( _get_conf_var $_potbase pot.level )" != "1" ]; then
+						_error "-P $_potbase : it has to be of level 1"
+						create-help
+						${EXIT} 1
+					fi
+				fi
+				if [ -z "$_base" ]; then
+					_base=$( _get_pot_base $_potbase )
+					if [ -z "$_base" ]; then
+						_error "-P $potbase has no base??"
+						${EXIT} 1
+					fi
+					_debug "-P $_potbase induced -b $_base"
+				fi
+				if ! _is_base "$_base" quiet ; then
+					_error "$_base is not a valid base"
+					create-help
+					${EXIT} 1
+				fi
+				;;
+			2)
+				if [ -z "$_potbase" ]; then
+					_error "level $_lvl pots need another pot as reference"
+					create-help
+					${EXIT} 1
+				fi
+				if [ $( _get_conf_var $_potbase pot.level ) -lt 1 ]; then
+					_error "-P $_potbase : it has to be at least of level 1"
+					create-help
+					${EXIT} 1
+				fi
+				if ! _is_pot $_potbase ; then
+					_error "-P $_potbase : is not a pot"
+					create-help
+					${EXIT} 1
+				fi
+				if [ -n "$_base" ]; then
+					if ! _is_base "$_base" quiet ; then
+						_error "$_base is not a valid base"
+						create-help
+						${EXIT} 1
+					fi
+					if [ "$( _get_pot_base $_potbase )" != "$_base" ]; then
+						_error "-b $_base and -P $_potbase are not compatible"
+						${EXIT} 1
+					fi
+				else
+					_base=$( _get_pot_base $_potbase )
+					if [ -z "$_base" ]; then
+						_error "-P $potbase has no base??"
+						${EXIT} 1
+					fi
+					if ! _is_base "$_base" quiet ; then
+						_error "$_base (induced by the pot $_potbase) is not a valid base"
+						create-help
+						${EXIT} 1
+					fi
+				fi
+				;;
+			*)
+				_error "level $_lvl is not supported"
+				${EXIT} 1
+				;;
+		esac
+	fi
 	if [ -z "$_pname" ]; then
 		_error "pot name is missing"
 		create-help
 		${EXIT} 1
 	fi
-	if _is_pot $_pname quiet ; then
+	if _is_pot "$_pname" quiet ; then
 		_error "pot $_pname already exists"
 		${EXIT} 1
 	fi
-	if [ "$_ipaddr" = "inherit" -a $_staticip = "YES" ]; then
+	if [ "$_ipaddr" = "inherit" ] && [ $_staticip = "YES" ]; then
 		_info "-s option is ignored if -i is inherit"
 		_staticip="NO"
 	fi
@@ -475,6 +552,7 @@ pot-create()
 	if _is_verbose ; then
 		_info "Option summary"
 		_info "pname : $_pname"
+		_info "type  : $_type"
 		_info "base  : $_base"
 		_info "lvl   : $_lvl"
 		_info "dns   : $_dns"
@@ -482,10 +560,10 @@ pot-create()
 		_info "ip    : $_ipaddr"
 		_info "ip alias : $_staticip"
 	fi
-	if ! _cj_zfs $_pname $_base $_lvl $_potbase ; then
+	if ! _cj_zfs "$_pname" "$_type" "$_lvl" "$_base" "$_potbase" ; then
 		${EXIT} 1
 	fi
-	if ! _cj_conf $_pname $_base $_ipaddr $_staticip $_lvl $_dns $_potbase ; then
+	if ! _cj_conf "$_pname" "$_base" "$_ipaddr" "$_staticip" "$_lvl" "$_dns" "$_type" "$_potbase" ; then
 		${EXIT} 1
 	fi
 	if [ $_flv_default = "YES" ]; then