+++ tests/create1.sh
@@ -168,9 +168,10 @@ test_pot_create_020()
 	assertEquals "_is_uid0 calls" "1" "$ISUID0_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
+	assertEquals "_cj_zfs arg2" "multi" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -178,7 +179,8 @@ test_pot_create_020()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 
 	setUp
@@ -189,9 +191,10 @@ test_pot_create_020()
 	assertEquals "_is_uid0 calls" "1" "$ISUID0_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
+	assertEquals "_cj_zfs arg2" "multi" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "test-pot" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "test-pot" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -199,7 +202,8 @@ test_pot_create_020()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "test-pot" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "test-pot" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -248,9 +252,9 @@ test_pot_create_040()
 	assertEquals "_is_uid0 calls" "1" "$ISUID0_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "2" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "test-pot" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "test-pot" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -258,7 +262,8 @@ test_pot_create_040()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "2" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "test-pot" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "test-pot" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 
 	setUp
@@ -269,9 +274,9 @@ test_pot_create_040()
 	assertEquals "_is_uid0 calls" "1" "$ISUID0_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "2" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "test-pot" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "test-pot" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -279,7 +284,8 @@ test_pot_create_040()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "2" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "test-pot" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "test-pot" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -306,9 +312,9 @@ test_pot_create_060()
 	assertEquals "_is_vnet_up calls" "0" "$ISVNETUP_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -316,7 +322,8 @@ test_pot_create_060()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -331,9 +338,9 @@ test_pot_create_061()
 	assertEquals "_is_vnet_up calls" "0" "$ISVNETUP_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -341,7 +348,8 @@ test_pot_create_061()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -356,9 +364,9 @@ test_pot_create_062()
 	assertEquals "_is_vnet_up calls" "1" "$ISVNETUP_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -366,7 +374,8 @@ test_pot_create_062()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -381,9 +390,9 @@ test_pot_create_063()
 	assertEquals "_is_vnet_up calls" "0" "$ISVNETUP_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -391,7 +400,8 @@ test_pot_create_063()
 	assertEquals "_cj_conf arg4" "YES" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "inherit" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -418,9 +428,9 @@ test_pot_create_081()
 	assertEquals "_is_vnet_up calls" "0" "$ISVNETUP_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -428,7 +438,8 @@ test_pot_create_081()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "pot" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -443,9 +454,9 @@ test_pot_create_082()
 	assertEquals "_is_vnet_up calls" "1" "$ISVNETUP_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -453,7 +464,8 @@ test_pot_create_082()
 	assertEquals "_cj_conf arg4" "NO" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "pot" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 
@@ -468,9 +480,9 @@ test_pot_create_083()
 	assertEquals "_is_vnet_up calls" "0" "$ISVNETUP_CALLS"
 	assertEquals "_cj_zfs calls" "1" "$CJZFS_CALLS"
 	assertEquals "_cj_zfs arg1" "new-pot" "$CJZFS_CALL1_ARG1"
-	assertEquals "_cj_zfs arg2" "11.1" "$CJZFS_CALL1_ARG2"
 	assertEquals "_cj_zfs arg3" "1" "$CJZFS_CALL1_ARG3"
-	assertEquals "_cj_zfs arg4" "" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg4" "11.1" "$CJZFS_CALL1_ARG4"
+	assertEquals "_cj_zfs arg5" "" "$CJZFS_CALL1_ARG5"
 	assertEquals "_cj_conf calls" "1" "$CJCONF_CALLS"
 	assertEquals "_cj_conf arg1" "new-pot" "$CJCONF_CALL1_ARG1"
 	assertEquals "_cj_conf arg2" "11.1" "$CJCONF_CALL1_ARG2"
@@ -478,17 +490,17 @@ test_pot_create_083()
 	assertEquals "_cj_conf arg4" "YES" "$CJCONF_CALL1_ARG4"
 	assertEquals "_cj_conf arg5" "1" "$CJCONF_CALL1_ARG5"
 	assertEquals "_cj_conf arg6" "pot" "$CJCONF_CALL1_ARG6"
-	assertEquals "_cj_conf arg7" "" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg7" "multi" "$CJCONF_CALL1_ARG7"
+	assertEquals "_cj_conf arg8" "" "$CJCONF_CALL1_ARG8"
 	assertEquals "_cj_flv calls" "1" "$CJFLV_CALLS"
 }
 setUp()
 {
 	common_setUp
 	CJZFS_CALLS=0
-	CJZFS_CALL1_ARG4=
+	CJZFS_CALL1_ARG5=
 	CJCONF_CALLS=0
-	CJCONF_CALL1_ARG6=
-	CJCONF_CALL1_ARG7=
+	CJCONF_CALL1_ARG8=
 	CJFLV_CALLS=0
 	HELP_CALLS=0
 	ISVNETAVAIL_CALLS=0