+++ tests/create3.sh
@@ -55,7 +55,7 @@ cat()
 test_cj_conf_001()
 {
 	# level 0
-	_cj_conf new-pot 11.1 inherit NO 0 inherit
+	_cj_conf new-pot 11.1 inherit NO 0 inherit multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/bases/11.1/usr.local /tmp/jails/new-pot/m/usr/local" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -73,7 +73,7 @@ test_cj_conf_001()
 
 test_cj_conf_002()
 {
-	_cj_conf new-pot 11.1 inherit NO 1 inherit
+	_cj_conf new-pot 11.1 inherit NO 1 inherit multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -95,7 +95,7 @@ test_cj_conf_003()
 	echo "zpot/bases/11.1 /tmp/jails/test-pot/m ro" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/usr.local /tmp/jails/test-pot/m/usr/local zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/custom /tmp/jails/test-pot/m/opt/custom zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
-	_cj_conf new-pot 11.1 inherit NO 1 inherit test-pot
+	_cj_conf new-pot 11.1 inherit NO 1 inherit multi test-pot
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -118,7 +118,7 @@ test_cj_conf_004()
 	echo "zpot/bases/11.1 /tmp/jails/test-pot/m ro" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/usr.local /tmp/jails/test-pot/m/usr/local zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
 	echo "zpot/jails/test-pot/custom /tmp/jails/test-pot/m/opt/custom zfs-remount" >> /tmp/jails/test-pot/conf/fscomp.conf
-	_cj_conf new-pot 11.1 inherit NO 2 inherit test-pot
+	_cj_conf new-pot 11.1 inherit NO 2 inherit multi test-pot
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/test-pot/usr.local /tmp/jails/new-pot/m/usr/local ro" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -137,7 +137,7 @@ test_cj_conf_004()
 
 test_cj_conf_005()
 {
-	_cj_conf new-pot 11.1 inherit NO 2 inherit test-pot-2
+	_cj_conf new-pot 11.1 inherit NO 2 inherit multi test-pot-2
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/test-pot/usr.local /tmp/jails/new-pot/m/usr/local ro" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -155,7 +155,7 @@ test_cj_conf_005()
 
 test_cj_conf_006()
 {
-	_cj_conf new-pot 11.1 10.1.2.3 NO 1 inherit
+	_cj_conf new-pot 11.1 10.1.2.3 NO 1 inherit multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -174,7 +174,7 @@ test_cj_conf_006()
 
 test_cj_conf_007()
 {
-	_cj_conf new-pot 11.1 inherit NO 1 pot
+	_cj_conf new-pot 11.1 inherit NO 1 pot multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -193,7 +193,7 @@ test_cj_conf_007()
 
 test_cj_conf_008()
 {
-	_cj_conf new-pot 11.1 10.1.2.3 NO 1 pot
+	_cj_conf new-pot 11.1 10.1.2.3 NO 1 pot multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -213,7 +213,7 @@ test_cj_conf_008()
 
 test_cj_conf_009()
 {
-	_cj_conf new-pot 11.1 10.1.2.3 YES 1 pot
+	_cj_conf new-pot 11.1 10.1.2.3 YES 1 pot multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"
@@ -233,7 +233,7 @@ test_cj_conf_009()
 
 test_cj_conf_020()
 {
-	_cj_conf new-pot 11.1 inherit YES 1 pot
+	_cj_conf new-pot 11.1 inherit YES 1 pot multi
 	assertEquals "return code" "0" "$?"
 	assertEquals "fscomp args1" "zpot/bases/11.1 /tmp/jails/new-pot/m ro" "$(sed '1!d' /tmp/jails/new-pot/conf/fscomp.conf)"
 	assertEquals "fscomp args2" "zpot/jails/new-pot/usr.local /tmp/jails/new-pot/m/usr/local zfs-remount" "$(sed '2!d' /tmp/jails/new-pot/conf/fscomp.conf)"