+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- create-base : add option -b, to provide a specific name to a base
+- create-base : add option -b, to provide a specific name to a base and support multiple bases with the same FreeBSD version
+- create: add support to single dataset pots.
+- set-cmd: add the command to manage the command line that starts the container
 
 ## [0.5.5] 2018-04-18
 ### Added